Keep the registry text helpers clear of Electron so their test runs in CI
The test mocked `electron` to import the module, which held when the
file ran first and failed with "Export named 'app' not found" once
another test had loaded the real module ahead of it. The `.reg` text
builder and the `reg query` parser take strings and return strings, so
they move into a file with no Electron import and the mock goes. Also
indexes the two query results explicitly, which the tuple from a
mapped array did not type as present.
