+++ packages/app/lib/windows-mail-client.ts
@@ -8,14 +8,15 @@ import { ms } from "@meru/shared/ms";
 import { app, shell } from "electron";
 import { serializeError } from "serialize-error";
 import { log } from "./log";
+import {
+  buildRegistration,
+  MAILTO_PROG_ID,
+  parseUserChoiceProgIds,
+} from "./windows-mail-client-registry";
 
 const execFile = promisify(childProcess.execFile);
 
-const MAILTO_PROG_ID = "Meru.mailto";
-
 const USER_CHOICE_KEY = String.raw`HKCU\Software\Microsoft\Windows\Shell\Associations\UrlAssociations\mailto`;
-
-const USER_CHOICE_SUBKEYS = ["UserChoiceLatest", "UserChoice"];
 
 /**
  * Windows 11 22H2 added Meru's own page under Default apps, which
@@ -41,77 +42,6 @@ function getRegExePath() {
  */
 function getExecutablePath() {
   return process.env.PORTABLE_EXECUTABLE_FILE ?? process.execPath;
-}
-
-function escapeRegValue(value: string) {
-  return `"${value.replace(/\\/g, "\\\\").replace(/"/g, '\\"')}"`;
-}
-
-export function buildRegistration(executablePath: string) {
-  const icon = escapeRegValue(`${executablePath},0`);
-
-  const openCommand = escapeRegValue(`"${executablePath}"`);
-
-  const mailtoCommand = escapeRegValue(`"${executablePath}" "%1"`);
-
-  // The leading `-` deletes `Software\Meru`, the second set of capabilities
-  // earlier installers wrote: left behind, Default apps lists two Merus
-  const registration = String.raw`Windows Registry Editor Version 5.00
-
-[-HKEY_CURRENT_USER\Software\Meru]
-
-[HKEY_CURRENT_USER\Software\Classes\Meru.mailto]
-@="Meru URL"
-"FriendlyTypeName"="Meru URL"
-"EditFlags"=dword:00000002
-
-[HKEY_CURRENT_USER\Software\Classes\Meru.mailto\DefaultIcon]
-@=${icon}
-
-[HKEY_CURRENT_USER\Software\Classes\Meru.mailto\shell\open\command]
-@=${mailtoCommand}
-
-[HKEY_CURRENT_USER\Software\Classes\mailto]
-"URL Protocol"=""
-
-[HKEY_CURRENT_USER\Software\Clients\Mail\Meru]
-@="Meru"
-"LocalizedString"="Meru"
-
-[HKEY_CURRENT_USER\Software\Clients\Mail\Meru\DefaultIcon]
-@=${icon}
-
-[HKEY_CURRENT_USER\Software\Clients\Mail\Meru\shell\open\command]
-@=${openCommand}
-
-[HKEY_CURRENT_USER\Software\Clients\Mail\Meru\Protocols\mailto]
-@="URL:MailTo Protocol"
-"URL Protocol"=""
-"EditFlags"=dword:00000002
-
-[HKEY_CURRENT_USER\Software\Clients\Mail\Meru\Protocols\mailto\DefaultIcon]
-@=${icon}
-
-[HKEY_CURRENT_USER\Software\Clients\Mail\Meru\Protocols\mailto\shell\open\command]
-@=${mailtoCommand}
-
-[HKEY_CURRENT_USER\Software\Clients\Mail\Meru\Capabilities]
-"ApplicationName"="Meru"
-"ApplicationDescription"="The Gmail experience you deserve"
-"ApplicationIcon"=${icon}
-
-[HKEY_CURRENT_USER\Software\Clients\Mail\Meru\Capabilities\StartMenu]
-"Mail"="Meru"
-
-[HKEY_CURRENT_USER\Software\Clients\Mail\Meru\Capabilities\URLAssociations]
-"mailto"="Meru.mailto"
-
-[HKEY_CURRENT_USER\Software\RegisteredApplications]
-"Meru"="Software\\Clients\\Mail\\Meru\\Capabilities"
-`;
-
-  // `reg.exe import` parses the file as Windows text
-  return registration.replace(/\n/g, "\r\n");
 }
 
 /**
@@ -141,17 +71,6 @@ export async function registerWindowsMailClient() {
   }
 }
 
-/**
- * Every `REG_SZ` value in a `reg query` dump except the hash. Windows 11 24H2
- * moved the choice from a `ProgId` value on `UserChoice` to a `ProgId` subkey
- * of `UserChoiceLatest`, so the value's name is not relied on, only its data.
- */
-export function parseUserChoiceProgIds(output: string) {
-  return [...output.matchAll(/^\s*(?!Hash\s)\S+\s+REG_SZ\s+(\S.*?)\s*$/gm)].map(
-    (match) => match[1],
-  );
-}
-
 async function queryUserChoiceProgIds(subKey: string) {
   try {
     const { stdout } = await execFile(
@@ -175,7 +94,10 @@ async function queryUserChoiceProgIds(subKey: string) {
  * it answers yes for a Meru that no mailto link reaches.
  */
 export async function isWindowsDefaultMailClient() {
-  const [latest, legacy] = await Promise.all(USER_CHOICE_SUBKEYS.map(queryUserChoiceProgIds));
+  const [latest, legacy] = await Promise.all([
+    queryUserChoiceProgIds("UserChoiceLatest"),
+    queryUserChoiceProgIds("UserChoice"),
+  ]);
 
   return (latest.length > 0 ? latest : legacy).includes(MAILTO_PROG_ID);
 }
